pyproject.toml
[project]
name = "prescription-validator"
version = "1.0.0"
description = "Medical prescription validation environment for OpenEnv — trains AI to prevent medication errors"
authors = [{name = "Suhel Mulla", email = "suhelmulla22@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["openenv", "healthcare", "medical", "prescription", "safety", "reinforcement-learning"]

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "websockets>=11.0.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
inference = [
    "openai>=1.0.0",
]

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["prescription_validator*", "server*", "models*", "client*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]